fix: move log_tool_usage decorator to helpers.py
- Moved decorator from registry.py to helpers.py for reusability
- Updated backup.py to import from helpers
- Fixed ImportError in backup tool registration
